requirement.txt
absl-py==2.1.0
aniso8601==10.0.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.2
dm-tree==0.1.8
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.48.1
fqdn==1.5.1
greenlet==3.1.1
h11==0.14.0
h5py==3.10.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ImageHash==4.3.1
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.10.0
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kaggle==1.6.12
keras==3.0.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
multimethod==1.11.2
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.97.1
opendatasets==0.1.22
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandas-profiling==3.1.0
pandocfilters==1.5.1
parso==0.8.3
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pynput==1.7.7
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
scikit-learn==1.5.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.27.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.38
sqlparse==0.4.4
stack-data==0.6.3
tangled-up-in-unicode==0.1.0
tenacity==8.2.3
terminado==0.18.1
tesseract==0.1.3
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.14.1
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.1
visions==0.7.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
WMI==1.5.1
wsproto==1.2.0
